Doc update: Improve README example.
It makes more sense to check the published state of a post when attempting to edit.

This change also updates the example to use Rails 4 idioms in the controller code.

Also, Ruby style guides seem to agree: 'The and and or keywords are banned. It's just not worth it. Always use && and || instead.
' - https://github.com/styleguide/ruby