Make Pundit#authorize a module method.
There are cases where service objects could make use of the authorize
method. Make it a module method so it can be called without mixing in
Pundit.

Have the Pundit#authorize instance method delegate to Pundit::authorize
module method.

Be sure the instance method continues to cache the policy if found.