Improve model policy template convenience
With an ApplicationPolicy set up to deny all, the first step to take
after generating a model policy is to override ApplicationPolicy
methods.  This patch adds stub methods to the model policy template to
make that step more convenient.  The `new?` and `edit?` methods are
excluded because the default ApplicationPolicy already defers their
behavior to `create?` and `update?`, respectively.