Add Suffix Configuration
Pundit uses a `SUFFIX` constant for locating Policy classes inferred
from objects passed to the authorization helpers. It is very helpful to
override this suffix in cases where the `Policy` keyword is not
preferred.

As an example, in an application where insurance policies are managed,
introducing the `Policy` keyword to the domain language can cause logic
to get confusing.

To allow applications to override this suffix without removing and
reassigning the constant, this commit introduces a configuration class
to the gem that allows developers to define an configuration block in an
initializer to override this value.

This could be expanded upon later by either adding things to the
configuration or dynamically defining the `policy` and `policy!` helpers
to instead use the configured suffix.

Appropriately this commit also updates the documentation to describe
configuring this suffix which remains `Policy` by default.