Defined Pundit::Error using compact style.
When defining the class with:

```
module Pundit
  class Error
  end
end
```

the `Error` class is mixed in to the controller when using
`include Pundit`. This can cause issues since `Error` is a
very common class name. Using the compact style:

```
class Pundit::Error
```

we avoid this, since it only creates the constant `Pundit::Error`.