Build against currently supported rubies
...and use the latest (2020) Travis build environment: focal

2.6 is the earliest currently supported Ruby version:

  https://www.ruby-lang.org/en/downloads/branches/

The focal environment doesn't seem to support openjdk 8. Here's the
failure from the build:

    install-jdk.sh 2020-06-02
    Expected feature release number in range of 9 to 17, but got: 8
    The command "~/bin/install-jdk.sh --target "/home/travis/openjdk8" --workspace "/home/travis/.cache/install-jdk" --feature "8" --license "GPL" --cacerts" failed and exited with 3 during .

It looks like JRuby 9.1 isn't under development: the last release was in
2018, and it's described as being "compatible with Ruby 2.3.x", which is
a very old version . It feels OK to drop it.

JRuby 9.2 says it's "compatible with Ruby 2.5.x", but it's the latest
JRuby version so it seems like something we should keep.