Use `scope.none` by default in ApplicationPolicy
A01:2021-Broken Access Control is the category with the most serious web
application security risk in 2021.

Using `scope.all` in the ApplicationPolicy template violates the
principle of least privilege or deny by default, where access should
only be granted for particular capabilities, roles, or users.

This change improves the security of the default ApplicationPolicy
template

Ref: https://owasp.org/Top10/A01_2021-Broken_Access_Control/