Prefer policy_cache if available to policy_finder
I'm unsure if there are other implications for this, but for the use
case of sharing context between the typical controller use (`authorize`)
vs the view typical use (`policy`) this adds value. Without this the
`authorize` and `policy` methods are still disconnected as far as I can
tell.

It looks like you were thoughtful about only choosing the else clause to
be stored in the cache, so I might be missing a consideration you were
making. But, I'd argue that if the client specifies a policy_class for
a given record then the goal is for that policy_class to apply to all
records through the pundit-based interactions. It was our intention at
least.